52 - FormatDateRangeFilter

+++ src/_11ty/filters/formatDate.js
@@ -2,7 +2,7 @@ module.exports = (date, locale = 'en-GB') => {
   // May need to consider `timezone` option (defaults to UTC)
   return new Intl.DateTimeFormat(locale, {
     year: 'numeric',
-    month: 'long',
+    month: 'short',
     day: 'numeric',
     // hour: 'numeric',
     // minute: 'numeric'
